refactor: centralize random string generation
